% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikimapR.R
\name{clear_wikimapia_api_keys}
\alias{clear_wikimapia_api_keys}
\title{Clear Wikimapia API keys}
\usage{
clear_wikimapia_api_keys()
}
\description{
Clears the API keys
}
\references{
mapdeck https://github.com/SymbolixAU/mapdeck code related to API key handling was used as a template.
}
