% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikimapR.R
\name{set_wikimapia_api_key}
\alias{set_wikimapia_api_key}
\title{Set Wikimapia API Key}
\usage{
set_wikimapia_api_key(api_key)
}
\arguments{
\item{api_key}{Wikimapia API key}
}
\description{
Sets an API key to make it available for all wikimapR functions calls. See details
}
\details{
Use \code{set_wikimapia_api_key} to make API keys available for all Wikimapia API
calls in a session so you don't have to keep specifying the \code{wm_api_key} argument
each time.
}
\references{
mapdeck https://github.com/SymbolixAU/mapdeck code related to API key handling was used as a template.
}
