% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikimapR.R
\docType{package}
\name{wikimapR}
\alias{wikimapR}
\alias{wikimapR-package}
\title{Import 'Wikimapia' Data as Simple Features via API}
\description{
Download and import of 'Wikimapia' data as 'sf'
objects.  'Wikimapia' data are extracted via the 'Wikimapia' API and
converted to `sf` objects retaining all recevied fields.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-kotov/wikimapR/}
  \item \url{http://www.ekotov.pro/wikimapR/}
  \item Report bugs at \url{https://github.com/e-kotov/wikimapR/issues}
}

}
\author{
Egor Kotov \email{kotov.egor@gmail.com}
}
