% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_get_by_id.R
\name{wm_get_by_id_single}
\alias{wm_get_by_id_single}
\title{Get one Wikimapia object by ID}
\usage{
wm_get_by_id_single(
  x,
  language = "ru",
  wm_api_key = "example",
  data_blocks = "main,geometry,edit,location,attached,photos,comments,translate"
)
}
\arguments{
\item{x}{a numeric or characted vector of length 1 with Wikimapia object ID to fetch. You can get a list of objects to fetch using \link[wikimapR]{wm_get_from_bbox}}

\item{language}{Wikimapia language to retrieve. This is specified language in ISO 639-1 format. Default language is 'ru'.}

\item{wm_api_key}{Your wikimapia API Key. If not specified, the default 'example' key is used, however it is limited to about 1 request per 30 seconds and is for testing purposes only.}

\item{data_blocks}{The comma-separated blocks of data you want to return: main, geometry, edit, location, attached, photos, comments, translate, similar_places, nearest_places, nearest_comments, nearest_streets, nearest_hotels. Default is "main,geometry,edit,location,attached,photos,comments,translate".}
}
\value{
`sf` object with nested details.
}
\description{
Retrieves single Wikimapia object by ID.
}
\details{
Uses [`place.getbyid`](http://wikimapia.org/api#placegetbyid) Wikimapia API function to retrieve full object details.
}
