% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdivide_bbox.R
\name{subdivide_bbox}
\alias{subdivide_bbox}
\title{Subdivide large bounding box into smaller ones}
\usage{
subdivide_bbox(x, bbox_cell_size = 0.045, return_bbox_or_sf = "bbox")
}
\arguments{
\item{x}{a numeric vector of length four, with xmin, ymin, xmax and ymax values; or a `bbox` object that is an output of \link[sf]{st_bbox}. The values must be in WGS 84 (EPSG 4326)}

\item{bbox_cell_size}{bounding box cell size is in degrees (need to somehow fix that to work with meters across the globe). Default of 0.045 is reasonably large, roughly equivalent to 2845x5010 meters.}

\item{return_bbox_or_sf}{object type to retun. If set to default "bbox", returns objects of type `bbox`, "sf" - returns `sf polygons` (good for plotting), "both" - returns a list of length 2 with both `bbox` and `sf` named accorgingly.}
}
\value{
an `sf` object grid with consequtive IDs. Also prints out a message with the number of bounding boxes and the average cell dimensions in meters.
}
\description{
Subdivide large bounding box into smaller ones.
}
